# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/bureau": {
      "get": {
        "description": "Page publique « Le bureau ». Ni adresse, ni identifiant de compte : publier les adresses de la promotion les livrerait aux robots collecteurs. Renvoie null tant qu’aucun mandat n’est actif.",
        "operationId": "BureauController_bureauPublic",
        "parameters": [],
        "responses": {
          "200": {
            "description": "La composition publiée, ou null si aucun mandat n’est actif.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BureauPublic"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides — le détail par champ est dans « champs ».",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, expiré ou invalide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "403": {
            "description": "Droits insuffisants pour cette action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "429": {
            "description": "Plafond de requêtes atteint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          }
        },
        "summary": "Composition du bureau en cours",
        "tags": [
          "Bureau COCFET"
        ]
      }
    },
    "/api/v1/bureau/postes": {
      "get": {
        "operationId": "BureauController_listerPostes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Le catalogue des postes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PosteBureau"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Données invalides — le détail par champ est dans « champs ».",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, expiré ou invalide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "403": {
            "description": "Droits insuffisants pour cette action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "429": {
            "description": "Plafond de requêtes atteint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les postes du bureau",
        "tags": [
          "Bureau COCFET"
        ]
      },
      "post": {
        "description": "Le catalogue est une donnée, pas une énumération : chaque mandat s’organise à sa façon, sans redéploiement.",
        "operationId": "BureauController_creerPoste",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreerPosteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Le poste créé.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosteBureau"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides — le détail par champ est dans « champs ».",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, expiré ou invalide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "403": {
            "description": "Droits insuffisants pour cette action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "409": {
            "description": "Ce nom de poste existe déjà.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "429": {
            "description": "Plafond de requêtes atteint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer un poste",
        "tags": [
          "Bureau COCFET"
        ]
      }
    },
    "/api/v1/bureau/postes/{id}": {
      "patch": {
        "operationId": "BureauController_mettreAJourPoste",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MettreAJourPosteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Le poste mis à jour.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosteBureau"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides — le détail par champ est dans « champs ».",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, expiré ou invalide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "403": {
            "description": "Droits insuffisants pour cette action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "404": {
            "description": "Poste inconnu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "409": {
            "description": "Ce nom de poste existe déjà.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "429": {
            "description": "Plafond de requêtes atteint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modifier un poste",
        "tags": [
          "Bureau COCFET"
        ]
      },
      "delete": {
        "description": "Refusé dès qu’un mandat l’a attribué : effacer le poste effacerait la trace de qui l’a occupé. Un poste devenu inutile se retire en ne le pourvoyant plus.",
        "operationId": "BureauController_supprimerPoste",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Poste supprimé."
          },
          "400": {
            "description": "Données invalides — le détail par champ est dans « champs ».",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, expiré ou invalide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "403": {
            "description": "Droits insuffisants pour cette action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "404": {
            "description": "Poste inconnu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "409": {
            "description": "Ce poste a déjà été occupé.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "429": {
            "description": "Plafond de requêtes atteint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer un poste",
        "tags": [
          "Bureau COCFET"
        ]
      }
    },
    "/api/v1/bureau/{generationId}/membres": {
      "get": {
        "operationId": "BureauController_listerMembres",
        "parameters": [
          {
            "name": "generationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Les postes pourvus pour ce mandat.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MembreBureau"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Données invalides — le détail par champ est dans « champs ».",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, expiré ou invalide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "403": {
            "description": "Droits insuffisants pour cette action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "404": {
            "description": "Génération inconnue.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "429": {
            "description": "Plafond de requêtes atteint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Composition d’un mandat",
        "tags": [
          "Bureau COCFET"
        ]
      },
      "post": {
        "description": "Le titulaire doit être un finissant de cette génération : le COCFET est le comité d’organisation de la cérémonie de fin d’étude, composé de ceux qui la vivent. C’est ainsi que le bureau sortant constitue le bureau entrant.",
        "operationId": "BureauController_affecter",
        "parameters": [
          {
            "name": "generationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AffecterMembreDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Le membre affecté.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembreBureau"
                }
              }
            }
          },
          "400": {
            "description": "Compte d’une autre promotion, ou adresse non confirmée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, expiré ou invalide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "403": {
            "description": "Droits insuffisants pour cette action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "404": {
            "description": "Génération, poste ou compte inconnu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "409": {
            "description": "Poste déjà occupé pour ce mandat.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "429": {
            "description": "Plafond de requêtes atteint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Désigner le titulaire d’un poste",
        "tags": [
          "Bureau COCFET"
        ]
      }
    },
    "/api/v1/bureau/membres/{id}": {
      "patch": {
        "operationId": "BureauController_mettreAJourMembre",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MettreAJourMembreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Le membre mis à jour.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembreBureau"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides — le détail par champ est dans « champs ».",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, expiré ou invalide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "403": {
            "description": "Droits insuffisants pour cette action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "404": {
            "description": "Membre inconnu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "429": {
            "description": "Plafond de requêtes atteint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modifier la présentation d’un membre",
        "tags": [
          "Bureau COCFET"
        ]
      },
      "delete": {
        "operationId": "BureauController_retirer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Membre retiré."
          },
          "400": {
            "description": "Données invalides — le détail par champ est dans « champs ».",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, expiré ou invalide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "403": {
            "description": "Droits insuffisants pour cette action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "404": {
            "description": "Membre inconnu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          },
          "429": {
            "description": "Plafond de requêtes atteint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErreurDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retirer un membre du bureau",
        "tags": [
          "Bureau COCFET"
        ]
      }
    },
...
  },
  "info": {
    "title": "COCFET API",
    "description": "API de la plateforme du Bureau des Finissants de l'UCAC-ICAM.\n\n### Authentification\n\nToute route exige un jeton, sauf mention contraire. Le jeton se\nprésente en en-tête `Authorization: Bearer <accessToken>`, et s'obtient\npar `POST /auth/connexion`.\n\nL'accès est de courte durée : quand il expire, `POST /auth/rafraichir`\nen délivre un nouveau. **Chaque rafraîchissement invalide le\nprécédent** — rejouer un ancien refresh token coupe toutes les\nsessions, le vol étant l’explication la plus probable.\n\n### Erreurs\n\nToutes les erreurs partagent la même forme, décrite par le schéma\n`ReponseErreurDto`. Le champ `code` est stable et destiné au code\nappelant ; `message` est destiné à l’affichage. Sur un 400, `champs`\ndétaille les erreurs de validation par champ.\n\n### Listes\n\nLes listes acceptent `page`, `limite` (plafonnée à 100), `tri` et\n`ordre`, et répondent `{ donnees, meta }`.\n\n### Fichiers\n\nLes fichiers ne sont jamais servis par une URL permanente. L'API\nrenvoie des **clés de stockage** ; une URL signée expirante s’obtient\npar `GET /fichiers/url-signee`. Ne stockez jamais une URL signée : elle\nserait périmée à la relecture.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ReponseErreurDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION",
            "description": "Code stable, destiné au code appelant. Contrairement au message, il ne change pas quand la formulation évolue."
          },
          "message": {
            "type": "string",
            "example": "Les données envoyées sont invalides.",
            "description": "Message destiné à l’affichage, en français."
          },
          "champs": {
            "type": "object",
            "description": "Erreurs de validation regroupées par champ. Présent uniquement sur un 400.",
            "example": {
              "email": [
                "email doit être une adresse valide"
              ]
            },
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "horodatage": {
            "type": "string",
            "example": "2027-08-12T18:00:00.000Z"
          },
          "chemin": {
            "type": "string",
            "example": "/api/v1/auth/connexion"
          },
          "identifiantCorrelation": {
            "type": "string",
            "example": "3f2a1b4c-5d6e-4f70-8a9b-0c1d2e3f4a5b",
            "description": "Identifiant à communiquer au support : il permet de retrouver la trace complète de l’erreur dans les journaux, sans que celle-ci soit exposée."
          }
        },
        "required": [
          "code",
          "message",
          "horodatage",
          "chemin",
          "identifiantCorrelation"
        ]
      },
...
      "BureauPublic": {
        "type": "object",
        "properties": {
          "annee": {
            "type": "number",
            "example": 2027
          },
          "nom": {
            "type": "string",
            "example": "ATLAS",
            "description": "Nom du bureau."
          },
          "logo": {
            "type": "object",
            "nullable": true,
            "description": "Logo désigné pour la plateforme."
          },
          "couleurPrimaire": {
            "type": "string",
            "example": "#0F172A"
          },
          "couleurSecondaire": {
            "type": "string",
            "example": "#D4AF37"
          },
          "membres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MembrePublic"
            }
          }
        },
        "required": [
          "annee",
          "nom",
          "logo",
          "couleurPrimaire",
          "couleurSecondaire",
          "membres"
        ]
      },
      "PosteBureau": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "nom": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "ordre": {
            "type": "number"
          },
          "estCle": {
            "type": "boolean"
          },
          "accordeAdministration": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "nom",
          "description",
          "ordre",
          "estCle",
          "accordeAdministration"
        ]
      },
      "CreerPosteDto": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string",
            "example": "Président"
          },
          "description": {
            "type": "string"
          },
          "ordre": {
            "type": "number",
            "example": 1,
            "description": "Ordre protocolaire. 1 = président."
          },
          "estCle": {
            "type": "boolean",
            "description": "Poste sans lequel un mandat n’a pas de sens : une génération dont les postes clés ne sont pas pourvus ne peut pas être activée."
          },
          "accordeAdministration": {
            "type": "boolean",
            "description": "Le titulaire administre la plateforme. C’est par ce drapeau que la passation d’administration se fait à chaque changement de mandat."
          }
        },
        "required": [
          "nom"
        ]
      },
      "MettreAJourPosteDto": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string",
            "example": "Président"
          },
          "description": {
            "type": "string"
          },
          "ordre": {
            "type": "number",
            "example": 1,
            "description": "Ordre protocolaire. 1 = président."
          },
          "estCle": {
            "type": "boolean",
            "description": "Poste sans lequel un mandat n’a pas de sens : une génération dont les postes clés ne sont pas pourvus ne peut pas être activée."
          },
          "accordeAdministration": {
            "type": "boolean",
            "description": "Le titulaire administre la plateforme. C’est par ce drapeau que la passation d’administration se fait à chaque changement de mandat."
          }
        }
      },
...
    }
  }
}